No need to use Char::Reader in PathReader
Given that we are only looking for the ascii characters '/' and '%', and
UTF-8 guarantees that these will always come as individual bytes in the
beginning of a character, we can simply traverse the string by bytes.
This avoids decoding UTF-8 and gives a small performance boost.

Benchmark times are updated as follows:

Before:
router 720.56k (  1.39µs) (± 1.16%)  1.33kB/op  fastest

After:
router 767.78k (  1.30µs) (± 3.48%)  1.33kB/op  fastest