Benchmark runs on github actions with PR commenting. (#69)
* Benchmark runs on github actions with PR commenting.

This github action will be run on PR's to compare benchmark times to the
base branch. After the comparison, it will add post a comment to the PR
with results. The comment contains the runtime information as well as
the comparison results, negative or positive.

The github action does require a `GH_TOKEN` action secret to be setup
first. The token needs the following permissions on this repo:

- Read access to metadata
- Read and Write access to code and pull requests

* Updating checkout from v2 to v4

* Explicitly targeting Crystal latest.

* Linting changes with Crystal 1.16.3